+++ tests/src/test_cli.cpp
@@ -414,11 +414,11 @@ TEST_F(CLI, override_handler)
     }
 }
 
-TEST_F(CLI, passwodr_mode)
+TEST_F(CLI, password_mode)
 {
     // initializer_list<const char*>
     {
-        EXPECT_STREQ("", twins::cli::passwordValue().cstr());
+        EXPECT_FALSE(twins::cli::passwordValue().hasPassword());
 
         twins::cli::passwordModeEnable(true);
         EXPECT_TRUE(twins::cli::passwordModeIsEnabled());
@@ -430,8 +430,11 @@ TEST_F(CLI, passwodr_mode)
     // initializer_list<const char*>
     {
         const twins::cli::Cmd commands[] = { { /* terminator */ } };
+        twins::SecurePassw pswQwert{"qwerty"};
+        twins::SecurePassw pswPo{"po"};
+        twins::SecurePassw pswLelum{"lelum"};
 
-        twins::cli::passwordSet({"qwerty", "po", "lelum"});
+        twins::cli::passwordSet({pswQwert, pswPo, pswLelum});
         twins::cli::passwordModeEnable(true);
         EXPECT_TRUE(twins::cli::passwordModeIsEnabled());
 
@@ -439,34 +442,35 @@ TEST_F(CLI, passwodr_mode)
         twins::cli::processInput("X\r");
         EXPECT_TRUE(twins::cli::checkAndExec(commands));
         EXPECT_FALSE(twins::cli::passwordModeIsEnabled());
-        EXPECT_STREQ("", twins::cli::passwordValue().cstr());
+        EXPECT_FALSE(twins::cli::passwordValue().hasPassword());
 
         // correct password
         twins::cli::processInput("qwerty\r");
         twins::cli::passwordModeEnable(true);
         EXPECT_TRUE(twins::cli::checkAndExec(commands));
         EXPECT_FALSE(twins::cli::passwordModeIsEnabled());
-        EXPECT_STREQ("qwerty", twins::cli::passwordValue().cstr());
+        EXPECT_TRUE(twins::cli::passwordValue() == "qwerty");
+        EXPECT_TRUE(twins::cli::passwordValue() == pswQwert);
     }
 
     // Vector<String> &&
-    {
-        const twins::cli::Cmd commands[] = { { /* terminator */ } };
+    // {
+    //     const twins::cli::Cmd commands[] = { { /* terminator */ } };
 
-        twins::Vector<twins::String> passwords = {"qwerty", "po", "lelum"};
-        twins::cli::passwordSet(std::move(passwords));
-        EXPECT_TRUE(passwords.size() == 0);
+    //     twins::Vector<twins::String> passwords = {"qwerty", "po", "lelum"};
+    //     twins::cli::passwordSet(std::move(passwords));
+    //     EXPECT_TRUE(passwords.size() == 0);
 
-        twins::cli::passwordModeEnable(true);
-        EXPECT_TRUE(twins::cli::passwordModeIsEnabled());
+    //     twins::cli::passwordModeEnable(true);
+    //     EXPECT_TRUE(twins::cli::passwordModeIsEnabled());
 
-        // correct password
-        twins::cli::processInput("lelum\r");
-        EXPECT_TRUE(twins::cli::checkAndExec(commands));
-        EXPECT_FALSE(twins::cli::passwordModeIsEnabled());
-        EXPECT_STREQ("lelum", twins::cli::passwordValue().cstr());
-        EXPECT_STREQ("", twins::cli::passwordValue().cstr());
-    }
+    //     // correct password
+    //     twins::cli::processInput("lelum\r");
+    //     EXPECT_TRUE(twins::cli::checkAndExec(commands));
+    //     EXPECT_FALSE(twins::cli::passwordModeIsEnabled());
+    //     EXPECT_STREQ("lelum", twins::cli::passwordValue().cstr());
+    //     EXPECT_STREQ("", twins::cli::passwordValue().cstr());
+    // }
 }
 
 #pragma GCC diagnostic pop // ignored "-Wunused-parameters"