+++ CHANGELOG.md
@@ -6,6 +6,8 @@ See the [keepachangelog.com description](https://keepachangelog.com/en/1.0.0/).
 
 * Added
   * twins::SecurePassw
+* Changed
+  * CLI uses SecurePassw instead of raw strings
 
 ## 0.26.0 - 2025-06-26